refactor builder api service to match existing patterns
