--- airbyte-webapp/src/services/connector-builder/ConnectorBuilderApiService.ts
@@ -1,157 +1,0 @@
-import { useQuery } from "react-query";
-
-import { useConfig } from "config";
-import { AirbyteRequestService } from "core/request/AirbyteRequestService";
-import {
-  StreamRead,
-  StreamReadRequestBody,
-  StreamReadRequestBodyConfig,
-  StreamReadRequestBodyManifest,
-  StreamsListRead,
-  StreamsListRequestBody,
-} from "core/request/ConnectorBuilderClient";
-import { useSuspenseQuery } from "services/connector/useSuspenseQuery";
-import { useDefaultRequestMiddlewares } from "services/useDefaultRequestMiddlewares";
-import { useInitService } from "services/useInitService";
-
-class ConnectorBuilderService extends AirbyteRequestService {
-  public readStream(body: StreamReadRequestBody): Promise<StreamRead> {
-    // TODO: uncomment this once there is a real API to call
-    // return readStream(body, this.requestOptions);
-    console.log("------------");
-    console.log(`Stream: ${body.stream}`);
-    console.log(`Connector manifest:\n${JSON.stringify(body.manifest)}`);
-    console.log(`Config:\n${JSON.stringify(body.config)}`);
-    return new Promise((resolve) => setTimeout(resolve, 200)).then(() => {
-      return {
-        slices: [
-          {
-            sliceDescriptor: { start: "Jan 1, 2022", end: "Jan 2, 2022" },
-            pages: [
-              {
-                airbyteMessages: [
-                  {
-                    type: "RECORD",
-                    record: {
-                      stream: body.stream,
-                      data: {
-                        id: "dp_123",
-                        object: body.stream,
-                        amount: 2000,
-                        balance_transaction: "txn_123",
-                      },
-                    },
-                  },
-                  {
-                    type: "STATE",
-                    state: {
-                      data: {
-                        timestamp: "2022-10-20T02:00:59Z",
-                      },
-                    },
-                  },
-                ],
-                request: {
-                  url: "https://api.com/path",
-                },
-                response: {
-                  status: 200,
-                },
-              },
-              {
-                airbyteMessages: [
-                  {
-                    type: "RECORD",
-                    record: {
-                      stream: body.stream,
-                      data: {
-                        id: "dp_123",
-                        object: body.stream,
-                        amount: 2000,
-                        balance_transaction: "txn_123",
-                      },
-                    },
-                  },
-                  {
-                    type: "STATE",
-                    state: {
-                      data: {
-                        timestamp: "2022-10-20T02:00:59Z",
-                      },
-                    },
-                  },
-                ],
-                request: {
-                  url: "https://api.com/path",
-                },
-                response: {
-                  status: 200,
-                },
-              },
-            ],
-          },
-        ],
-      };
-    });
-  }
-
-  public listStreams(body: StreamsListRequestBody): Promise<StreamsListRead> {
-    // TODO: uncomment this once there is a real API to call
-    // return listStreams(body, this.requestOptions);
-    console.log(`Received listStreams body: ${JSON.stringify(body)}`);
-    return new Promise((resolve) => setTimeout(resolve, 200)).then(() => {
-      return {
-        streams: [
-          {
-            name: "disputes",
-            url: "https://api.com/disputes",
-          },
-          {
-            name: "transactions",
-            url: "https://api.com/transactions",
-          },
-          {
-            name: "users",
-            url: "https://api.com/users",
-          },
-        ],
-      };
-    });
-  }
-}
-
-const connectorBuilderKeys = {
-  all: ["connectorBuilder"] as const,
-  read: (
-    streamName: string,
-    manifest: StreamReadRequestBodyManifest,
-    config: StreamReadRequestBodyConfig
-  ) => [...connectorBuilderKeys.all, "read", { streamName, manifest, config }] as const,
-  list: (manifest: StreamReadRequestBodyManifest) =>
-    [...connectorBuilderKeys.all, "list", { manifest }] as const,
-};
-
-function useConnectorBuilderService() {
-  const config = useConfig();
-  const middlewares = useDefaultRequestMiddlewares();
-  return useInitService(
-    () => new ConnectorBuilderService(config.connectorBuilderApiUrl, middlewares),
-    [config.connectorBuilderApiUrl, middlewares]
-  );
-}
-
-export const useReadStream = (params: StreamReadRequestBody) => {
-  const service = useConnectorBuilderService();
-
-  return useQuery(
-    connectorBuilderKeys.read(params.stream, params.manifest, params.config),
-    () => service.readStream(params),
-    { refetchOnWindowFocus: false, enabled: false }
-  );
-};
-
-export const useListStreams = (params: StreamsListRequestBody) => {
-  const service = useConnectorBuilderService();
-
-  return useSuspenseQuery(connectorBuilderKeys.list(params.manifest), () => service.listStreams(params));
-};
